Fix clippy

+++ parcel_sourcemap/src/lib.rs
@@ -194,7 +194,7 @@ impl SourceMap {
                     previous_source = original_source;
 
                     let original_line = original.original_line as i64;
-                    vlq::encode((original_line - previous_original_line) as i64, output)?;
+                    vlq::encode(original_line - previous_original_line, output)?;
                     previous_original_line = original_line;
 
                     let original_column = original.original_column as i64;
