+++ parcel_sourcemap_wasm/src/lib.rs
@@ -125,7 +125,7 @@ impl SourceMap {
         for mapping in self.map.get_mappings().iter() {
             mappings.push(MappingResult {
                 generated: PositionResult {
-                    line: (mapping.generated_line + 1) as u32,
+                    line: mapping.generated_line + 1,
                     column: mapping.generated_column,
                 },
                 original: mapping.original.map(|p| PositionResult {
